requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
PyMySQL==1.1.0
bcrypt==4.1.2
cryptography==41.0.7
email-validator==2.1.0